requirements.txt

streamlit
pandas
numpy
requests
lxml
beautifulsoup4
pyyaml
python-dateutil
tqdm
openpyxl
